requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
sarvamai
langchain-google-genai
streamlit
requests
langchain
langchain-community